Extract checkDatabaseMatches test helper
